Create MainActivity.kt
Including an alarm feature that activates when the user pressed the "Discover Devices" button. Once the button is pressed, the timer will be run for 120 seconds (because the device will be discovering for 120 seconds), and during the 120 seconds, the user cannot press any other button until the discovering process (120 seconds) has been completed. After the 120 seconds, this page will be refreshed.